integrate dynamic route
