+++ packages/studiocms/src/routes/api/auth/[provider]/_effects/auth0.ts
@@ -3,57 +3,76 @@ import { authEnvCheck } from 'studiocms:auth/utils/authEnvCheck';
 import config, { authConfig } from 'studiocms:config';
 import { StudioCMSRoutes } from 'studiocms:lib';
 import { SDKCore } from 'studiocms:sdk';
-import { generateState } from 'arctic';
-import { GitHub } from 'arctic';
+import { generateCodeVerifier, generateState } from 'arctic';
+import { Auth0 } from 'arctic';
 import type { APIContext } from 'astro';
 import { Effect } from 'effect';
-import { genLogger } from '../../../../lib/effects/index.js';
+import { genLogger } from '../../../../../lib/effects/index.js';
 
-const {
-	GITHUB: { CLIENT_ID = '', CLIENT_SECRET = '', REDIRECT_URI = null },
+export const {
+	AUTH0: { CLIENT_ID = '', CLIENT_SECRET = '', DOMAIN, REDIRECT_URI = '' },
 } = await authEnvCheck(authConfig.providers);
 
-export const ProviderID = 'github';
-export const ProviderCookieName = 'github_oauth_state';
+export const getClientDomain = () => {
+	const cleanDomainSlash = DOMAIN ? DOMAIN.replace(/^\//, '') : '';
 
-export const github = new GitHub(CLIENT_ID, CLIENT_SECRET, REDIRECT_URI);
+	const NoHttpDomain = cleanDomainSlash.replace(/http:\/\//, '').replace(/https:\/\//, '');
 
-export interface GitHubUser {
-	id: number;
-	html_url: string;
-	login: string;
-	avatar_url: string;
+	return `https://${NoHttpDomain}`;
+};
+
+export const ProviderID = 'auth0';
+export const ProviderCookieName = 'auth0_oauth_state';
+export const ProviderCodeVerifier = 'auth0_oauth_code_verifier';
+
+export const auth0 = new Auth0(getClientDomain(), CLIENT_ID, CLIENT_SECRET, REDIRECT_URI);
+
+export interface Auth0User {
+	sub: string;
 	name: string;
-	blog: string;
 	email: string;
+	picture: string;
+	nickname: string;
 }
 
-export class GitHubOAuthAPI extends Effect.Service<GitHubOAuthAPI>()('GitHubOAuthAPI', {
-	effect: genLogger('studiocms/routes/api/auth/github/effect')(function* () {
+export class Auth0OAuthAPI extends Effect.Service<Auth0OAuthAPI>()('Auth0OAuthAPI', {
+	effect: genLogger('studiocms/routes/api/auth/auth0/effect')(function* () {
 		const sessionHelper = yield* Session;
 		const sdk = yield* SDKCore;
 		const verifyEmail = yield* VerifyEmail;
 		const userLib = yield* User;
 
 		const initSession = (context: APIContext) =>
-			genLogger('studiocms/routes/api/auth/github/effect.initSession')(function* () {
+			genLogger('studiocms/routes/api/auth/auth0/effect.initSession')(function* () {
 				const state = generateState();
 
-				const scopes = ['user:email', 'repo'];
+				const codeVerifier = generateCodeVerifier();
 
-				const url = github.createAuthorizationURL(state, scopes);
+				const scopes = ['openid', 'profile', 'email'];
+
+				const url = auth0.createAuthorizationURL(state, codeVerifier, scopes);
 
 				yield* sessionHelper.setOAuthSessionTokenCookie(context, ProviderCookieName, state);
+
+				yield* sessionHelper.setOAuthSessionTokenCookie(
+					context,
+					ProviderCodeVerifier,
+					codeVerifier
+				);
 
 				return context.redirect(url.toString());
 			});
 
-		const validateAuthCode = (code: string) =>
-			genLogger('studiocms/routes/api/auth/github/effect.validateAuthCode')(function* () {
-				const tokens = yield* Effect.tryPromise(() => github.validateAuthorizationCode(code));
+		const validateAuthCode = (code: string, codeVerifier: string) =>
+			genLogger('studiocms/routes/api/auth/auth0/effect.validateAuthCode')(function* () {
+				const tokens = yield* Effect.tryPromise(() =>
+					auth0.validateAuthorizationCode(code, codeVerifier)
+				);
+
+				const CLIENT_DOMAIN = getClientDomain();
 
 				const response = yield* Effect.tryPromise(() =>
-					fetch('https://api.github.com/user', {
+					fetch(`${CLIENT_DOMAIN}/userinfo`, {
 						headers: {
 							Authorization: `Bearer ${tokens.accessToken}`,
 						},
@@ -64,30 +83,31 @@ export class GitHubOAuthAPI extends Effect.Service<GitHubOAuthAPI>()('GitHubOAut
 					yield* Effect.fail(new Error('Failed Authorization Check'));
 				}
 
-				const resData: GitHubUser = yield* Effect.tryPromise(() => response.json());
+				const resData: Auth0User = yield* Effect.tryPromise(() => response.json());
 
 				return resData;
 			});
 
 		const initCallback = (context: APIContext) =>
-			genLogger('studiocms/routes/api/auth/github/effect.initCallback')(function* () {
+			genLogger('studiocms/routes/api/auth/auth0/effect.initCallback')(function* () {
 				const { url, cookies, redirect } = context;
 
 				const code = url.searchParams.get('code');
 				const state = url.searchParams.get('state');
+				const codeVerifier = cookies.get(ProviderCodeVerifier)?.value ?? null;
 				const storedState = cookies.get(ProviderCookieName)?.value ?? null;
 
-				if (!code || !state || !storedState || state !== storedState) {
+				if (!code || !storedState || !codeVerifier || state !== storedState) {
 					return redirect(StudioCMSRoutes.authLinks.loginURL);
 				}
 
-				const githubUser = yield* validateAuthCode(code);
+				const auth0User = yield* validateAuthCode(code, codeVerifier);
 
-				const { id: githubUserId, login: githubUsername } = githubUser;
+				const { sub: auth0UserId, name: auth0Username } = auth0User;
 
 				const existingOAuthAccount = yield* sdk.AUTH.oAuth.searchProvidersForId(
 					ProviderID,
-					`${githubUserId}`
+					auth0UserId
 				);
 
 				if (existingOAuthAccount) {
@@ -121,7 +141,7 @@ export class GitHubOAuthAPI extends Effect.Service<GitHubOAuthAPI>()('GitHubOAut
 						yield* sdk.AUTH.oAuth.create({
 							userId: existingUser.id,
 							provider: ProviderID,
-							providerUserId: `${githubUserId}`,
+							providerUserId: auth0UserId,
 						});
 
 						const isEmailAccountVerified = yield* verifyEmail.isEmailVerified(existingUser);
@@ -143,14 +163,13 @@ export class GitHubOAuthAPI extends Effect.Service<GitHubOAuthAPI>()('GitHubOAut
 					{
 						// @ts-expect-error drizzle broke the id variable...
 						id: crypto.randomUUID(),
-						username: githubUsername,
-						email: githubUser.email,
-						name: githubUser.name,
-						avatar: githubUser.avatar_url,
+						username: auth0Username,
+						name: auth0User.name,
+						email: auth0User.email,
+						avatar: auth0User.picture,
 						createdAt: new Date(),
-						url: githubUser.blog,
 					},
-					{ provider: ProviderID, providerUserId: `${githubUserId}` }
+					{ provider: ProviderID, providerUserId: auth0UserId }
 				);
 
 				if ('error' in newUser) {
